Breaking VPN logic in logical pieces inside their respective module
