+++ ios/brave-ios/Sources/BraveVPN/Components/Region/BraveVPNRegionLoadingIndicator.swift
